complete DAL

+++ MyMiniLedger.DAL/Models/PositionModel.cs
@@ -13,9 +13,9 @@ namespace MyMiniLedger.DAL.Models
         public DateTime OpenDate { get; set; }
         public DateTime CloseDate { get; set; }
         public int KindId { get; set; }
-        public decimal Income { get; set; }
-        public decimal Expense { get; set; }
-        public decimal Saldo { get; set; }
+        public double Income { get; set; }
+        public double Expense { get; set; }
+        public double Saldo { get; set; }
         public int CoinId { get; set; }
         public int StatusId { get; set; }
         public string Tag { get; set; }
